Fix Buffer::create_shm_buffer — signed integer overflow in stride computation
-resolves https://github.com/jwinarske/waypp/issues/29

+++ examples/view_manager/view_wayland.cc
@@ -177,7 +177,7 @@ void ViewWayland::draw_frame(void* data, const uint32_t /* time */) {
       static_cast<uint32_t>(window->get_width()),
       static_cast<uint32_t>(window->get_height()), 8,
       static_cast<uint32_t*>(buffer->get_shm_data()),
-      static_cast<std::size_t>(buffer->get_size()));
+      buffer->get_size());
 
   wl_surface_attach(window->get_surface(), buffer->get_wl_buffer(), 0, 0);
   wl_surface_damage(window->get_surface(), 0, 0, window->get_width(),
